Remove unit tests and write parser tests, add AF_EMPTY magic word
This commit removes the previously written unit tests and replaces them with parser tests. The unit tests were not very useful, because they failed to test the interaction with the other components, such as the ArgumentPreprocessor. Some tests were also hard to implement as unit tests, because they had close interaction with the parser (such as #af_foreach). I therefore chose to test the parser functions with parser tests instead. For the other components, unit tests may still be useful to implement.