Test ArrayFunctions.ServiceWiring.php
Test ArrayFunctions.ServiceWiring.php using tests copied from
CentralAuth. Because phpunit does not support marking a file a
covered, the ServiceWiring file is ignored for code coverage as
the tests fully cover the file.

Add TestAutoloadNamespaces to extension.json which will autoload
phpunit test files for this extension.

Also update packages in package.json as the version of grunt
specified is vulnerable.

Depends-On: I164b0b38362f5a77bd38ef4f2453f00fc116bd48
Change-Id: I3abfa9311ca1a9ca663a114b06fa9d86f2b356bc