Remove unnecessary read check from #af_template
This commit removes an unnecessary read check from #af_template. This check used to check whether the user was allowed to read the page they were trying to transclude. Normal template transclusion does not perform this check either, so to make it more similar to regular template transclusion, the check has been removed.

Change-Id: I143401c19643bfdfd839a56381c101a7e67d8ca6