Improve handling of whitespace and add #af_trim
In this change, the handling of whitespace for arrays is improved. Similar to what is done for strings already by MediaWiki, whitespace is now trimmed from strings wrapped in arrays as well. Since handling of whitespace is already very implicit in MediaWiki, this should prevent users from accidentally inserting whitespace in their arrays.

Bug: T355929
Change-Id: I59510d22cae6b0cd982fbcd025a2a4463564a826