Add ZLIB compression using `gzdeflate` and `gzinflate`
This patch adds ZLIB compression to returned arrays to make them smaller in size. This makes it harder to reach the $wgMaxArticleSize limit. The arrays are first encoded into JSON (which is more space-efficient than serialize), then compressed if possible and finally they are base64-encoded. Encoding the compression result is necessary, because MediaWiki does not correctly handle the resulting binary.

This patch also fixes/removes a number of tests that made assertions about the opaque encoding.

Change-Id: If4e27a104c8c6de20690e791ee9f8dc1c594c310