Rename #af_wildcard to #af_group
This merge request renames #af_wildcard to #af_group, and adds the old #af_wildcard name as an alias. The function should be renamed becuse #af_wildcard is undescriptive, and even misleading, as it doesn't behave exactly as indexing with a wildcard would. #af_group describes the behaviour of the parser function better.

Change-Id: I5cdb8cf942620ca7b8b066fcd6aaaded4eb81392