Add #af_put parser function and deprecate #af_set
This merge request adds the #af_put parser function as a replace of the #af_set parser function, deprecating the latter. The #af_set parser function is the only parser function which takes an array as its second argument, as opposed to its first argument. The #af_put parser function swaps this to be in line with the other parser functions. In order to be backwards compatible, the old #af_set parser function remains unchanged.

Change-Id: I0dec6f3eadb07fe8bdebcf231a3b1591472781ed